Ported to 1.13.2

+++ src/main/java/net/geforcemods/rologia/os/gui/screens/input/InputManager.java
@@ -17,7 +17,7 @@ public class InputManager {
 		OS = os;
 	}
 
-	public void handleMouseClick(int mouseX, int mouseY, int mouseButtonClicked) {
+	public void handleMouseClick(double mouseX, double mouseY, int mouseButtonClicked) {
 		boolean clickedOnComponent = false;
 		Screen screen = OS.getCurrentScreen();
